requirements.txt
torch
numpy
Pillow
tqdm
matplotlib
scikit-learn
git+https://github.com/openai/CLIP.git